requirements.txt
websockets
flask
flask_sqlalchemy
pdfplumber
python-docx
pytesseract
selenium
undetected-chromedriver
fake-useragent
setuptools
wheel